+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to `ixudra/curl` will be documented in this file
 
+## 6.21.0 - 2020-09-22
+### Added
+- withAuthorization utility method
+- withBearer utility method
+
 ## 6.20.0 - 2020-08-07
 ### Fixed
 - Data is now passed as POST parameters instead of GET for `DELETE` REQUESTS